contributors/ttwf/betravis/submitted/style-sheet-interfaces-001.html

<!DOCTYPE html>
<html>
 <head>
  <title>CSS Test: CSSOM StyleSheet and StyleSheetList</title>
  <link rel="author" title="Bear Travis" href="mailto:betravis@adobe.com">
  <link rel="help" href="http://www.w3.org/TR/DOM-Level-2-Style/stylesheets.html#StyleSheets-fundamental">
  <link rel="help" href="http://www.w3.org/TR/DOM-Level-2-Style/stylesheets.html#StyleSheets-StyleSheet-DocumentStyle">
  <meta name="flags" content="dom">
  <script src="/resources/testharness.js" type="text/javascript"></script>
  <script src="/resources/testharnessreport.js" type="text/javascript"></script>
  <style>
    * { margin: 0; padding: 0; }
    .blue { background-color: green; }
  </style>
  <style id="styleElement" type="text/css" media="all" title="internal style sheet">
    .blue { background-color: blue; }
  </style>
 </head>
 <body>
  <noscript>Test not run - javascript required.</noscript>
  <div id="blue-div" class="blue"></div>
  <div id="log"></div>
  <script type="text/javascript">
    test(function() {
        assert_own_property(document, "styleSheets");
        assert_true(document.styleSheets instanceof StyleSheetList, "styleSheets property is a StyleSheetList");
    }, "document.styleSheets",
    { assert: ["document.styleSheets exists", "document.styleSheets is a StyleSheetList"] });

    var styleSheets = document.styleSheets;

    test(function() {
        assert_own_property(styleSheets, "length");
        assert_readonly(styleSheets, "length");
        assert_inherits(styleSheets, "item");
    }, "StyleSheetList properties",
    { assert: ["length property exists", "length property is read only", "item property exists"] });

    test(function() {
        assert_true(styleSheets.length > 1);
        assert_true(styleSheets.item(1) instanceof StyleSheet);
    }, "StyleSheetList property values",
    { assert: "document.styleSheets contains at least one StyleSheet" });

    var styleSheet = styleSheets.item(1);

    test(function() {
        assert_own_property(styleSheet, "type");
        assert_own_property(styleSheet, "disabled");
        assert_own_property(styleSheet, "ownerNode");
        assert_own_property(styleSheet, "parentStyleSheet");
        assert_own_property(styleSheet, "href");
        assert_own_property(styleSheet, "title");
        assert_own_property(styleSheet, "media");

        assert_readonly(styleSheet, "type");
        assert_readonly(styleSheet, "ownerNode");
        assert_readonly(styleSheet, "parentStyleSheet");
        assert_readonly(styleSheet, "href");
        assert_readonly(styleSheet, "title");
        assert_readonly(styleSheet, "media");
    }, "StyleSheet properties",
    { assert: [ "type, disabled, ownerNode, parentStyleSheet, href, title, and media properties exist on StyleSheet",
                "type, ownerNode, parentStyleSheet, href, title, media properties are read only" ] });

    test(function() {
        assert_equals(styleSheet.type, "text/css");
        assert_equals(styleSheet.disabled, false);
        assert_equals(styleSheet.ownerNode, document.getElementById("styleElement"));
        assert_equals(styleSheet.href, null);
        assert_equals(styleSheet.title, "internal style sheet");
        assert_equals(styleSheet.media.item(0), "all");
    }, "StyleSheet property values",
    { assert: "StyleSheet initial property values are correct" });

    var blueDiv = document.getElementById("blue-div");

    test(function() {
        assert_equals(getComputedStyle(blueDiv).backgroundColor, "rgb(0, 0, 255)");
        styleSheet.media.mediaText = "none";
        assert_equals(getComputedStyle(blueDiv).backgroundColor, "rgb(0, 128, 0)");
        styleSheet.media.mediaText = "all";
    }, "StyleSheet media value",
    { assert: "Setting media to 'none' disables a StyleSheet" });

    test(function() {        
        styleSheet.disabled = true;
        assert_equals(getComputedStyle(blueDiv).backgroundColor, "rgb(0, 128, 0)");
    }, "StyleSheet disabled value",
    { assert: "Setting disabled to 'true' disables a StyleSheet"});
  </script>
 </body>
</html>
